core: extract search introspection builders into core/search (modes-report, tune-recommendations, graph-signals stats)
Behavior-preserving move so the upcoming search_modes / search_stats /
search_tune MCP ops render the same reports as the CLI: buildModesReport +
KNOB_DESCRIPTIONS → core/search/modes-report.ts, the tune recommendation
builder → core/search/tune-recommendations.ts (strictly read-only — the apply
lane stays in the command file per CDX-21), readGraphSignalsStats →
core/search/telemetry.ts. src/commands/search.ts keeps arg parsing, text
rendering, the modes reset lane, and the apply/revert lane; all JSON output
shapes unchanged.
